Reformat proposal list
